+++ src/pages/arbeidsgiver/[narmestelederid]/[oppfolgingsdialogId]/arbeidsoppgaver.tsx
@@ -1,14 +1,14 @@
 import { NextPage } from "next";
 import React from "react";
-import {
-  OppfolgingsplanPageAG,
-  Page,
-} from "../../../../components/blocks/wrappers/OppfolgingsplanPageAG";
 import { LagredeArbeidsoppgaver } from "../../../../components/arbeidsoppgaver/LagredeArbeidsoppgaver";
 import { beskyttetSideUtenProps } from "../../../../auth/beskyttetSide";
 import { NyArbeidsoppgaveAG } from "../../../../components/arbeidsoppgaver/NyArbeidsoppgaveAG";
 import { useAktivPlanAG } from "../../../../api/queries/arbeidsgiver/oppfolgingsplanerQueriesAG";
 import { useInnloggetFnr } from "../../../../api/queries/oppfolgingsplan/oppfolgingsplanQueries";
+import {
+  OppfolgingsplanPageAG,
+  Page,
+} from "../../../../components/blocks/wrappers/oppfolgingsplanpageag/OppfolgingsplanPageAG";
 
 const Arbeidsoppgaver: NextPage = () => {
   const aktivPlan = useAktivPlanAG();
