+++ src/pages/arbeidsgiver/[narmestelederid]/[oppfolgingsdialogId]/seplanen.tsx
@@ -1,13 +1,13 @@
 import { NextPage } from "next";
-import {
-  OppfolgingsplanPageAG,
-  Page,
-} from "../../../../components/blocks/wrappers/OppfolgingsplanPageAG";
 import { OppfolgingsplanOversikt } from "../../../../components/seplanen/OppfolgingsplanOversikt";
 import { beskyttetSideUtenProps } from "../../../../auth/beskyttetSide";
 import { SendTilGodkjenningAG } from "../../../../components/seplanen/sendtilgodkjenning/SendTilGodkjenningAG";
 import { useAktivPlanAG } from "../../../../api/queries/arbeidsgiver/oppfolgingsplanerQueriesAG";
 import { GodkjennEgenPlanAG } from "../../../../components/seplanen/sendtilgodkjenning/GodkjennEgenPlanAG";
+import {
+  OppfolgingsplanPageAG,
+  Page,
+} from "../../../../components/blocks/wrappers/oppfolgingsplanpageag/OppfolgingsplanPageAG";
 
 const Seplanen: NextPage = () => {
   const aktivPlan = useAktivPlanAG();
